chore(DataStore): Minor refactor syncDeletion(s) and move queryAndDeleteTransaction
